requirements.txt
# django 
Django
djangorestframework
django-filter
drf-extensions
requests
markdown

# prod wsgi server
gunicorn

# postgres driver
psycopg2-binary

# opencv stuff
pillow
opencv-contrib-python
opencv-python-headless

# background jobs
redis
celery[redis]

# file bucket
google-cloud-storage